+++ app/src/main/java/eu/laramartin/weather/WeatherView.java
@@ -4,6 +4,6 @@ package eu.laramartin.weather;
  * Created by Lara on 21/10/2016.
  */
 public interface WeatherView {
-    void displayTemp(double temperature);
+    void displayTemp(int temperature);
     void displayDescription(String description);
 }
